+++ Sources/Migration/Migrator.swift
@@ -29,6 +29,10 @@
 // Usually reusing the same NSMigrationManager for multiple mapping models works fine... until it doesn't!
 // (in particular if the model has entities with "particular" rules, i.e. relationships with min and max set with custom values).
 // In these cases, one of the migrations could fail (mostly due to validation errors) unless we use different NSMigrationManager instance.
+//
+// In a subclass of NSPersistentStore, you can override this to provide a custom migration manager subclass
+// (for example, to take advantage of store-specific functionality to improve migration performance).
+// https://developer.apple.com/documentation/coredata/nspersistentstore/1506361-migrationmanagerclass
 
 import CoreData
 import os.log
@@ -44,7 +48,7 @@ final class Migrator<Version: ModelVersion>: NSObject, ProgressReporting {
     var progress = Progress(totalUnitCount: 1)
     return progress
   }()
-  
+
   /// Enable log.
   public var enableLog: Bool = false {
     didSet {
@@ -55,18 +59,18 @@ final class Migrator<Version: ModelVersion>: NSObject, ProgressReporting {
       }
     }
   }
-  
+
   private var log: OSLog = .disabled
-  
+
   /// Source description used as starting point for the migration steps.
   let sourceStoreDescription: NSPersistentStoreDescription
-  
+
   /// Desitnation description used as final point for the migrations steps.
   let destinationStoreDescription: NSPersistentStoreDescription
-  
+
   /// `Version` to which the database needs to be migrated.
   let targetVersion: Version
-  
+
   /// Creates a `Migrator` instance to handle a multi step migration to a given `Version`
   /// - Parameters:
   ///   - sourceStoreDescription: Initial persistent store description.
@@ -81,22 +85,23 @@ final class Migrator<Version: ModelVersion>: NSObject, ProgressReporting {
     super.init()
     _ = progress // lazy init for implicit progress support
   }
-  
+
   /// Creates a `Migrator` instance to handle a multi step migration at `targetStoreDescription` to a given `Version`.
   public convenience init(targetStoreDescription: NSPersistentStoreDescription, targetVersion: Version) {
     self.init(sourceStoreDescription: targetStoreDescription, destinationStoreDescription: targetStoreDescription, targetVersion: targetVersion)
   }
-  
+
   /// Migrates the store to a given `Version`, performing a WAL checkpoint if opted in.
   /// - Parameters:
   ///   - targetVersion: The final `Version`.
-  ///   - enableWALCheckpoint: Wheter or not a WAL checkpoint needs to be done. A dead lock can occur if a NSPersistentStore with a different journaling mode is currently active and using the database file.
+  ///   - enableWALCheckpoint: Wheter or not a WAL checkpoint needs to be done.
+  ///   A dead lock can occur if a NSPersistentStore with a different journaling mode is currently active and using the database file.
   ///   - managerProvider: Closure to provide a custom `NSMigrationManager` instance.
   /// - Throws: It throws an error if the migration fails.
   public func migrate(enableWALCheckpoint: Bool = false, managerProvider: ((Metadata) -> NSMigrationManager)? = nil) throws {
     guard let sourceURL = sourceStoreDescription.url else { fatalError("Source NSPersistentStoreDescription requires a URL.") }
     guard let destinationURL = destinationStoreDescription.url else { fatalError("Destination NSPersistentStoreDescription requires a URL.") }
-    
+
     try migrateStore(from: sourceURL,
                      sourceOptions: sourceStoreDescription.options,
                      to: destinationURL,
@@ -120,15 +125,15 @@ extension Migrator {
     let start = DispatchTime.now()
     guard let sourceVersion = try Version(persistentStoreURL: sourceURL) else {
       let message = "A ModelVersion could not be found for the initial store at: \(sourceURL)."
-      os_log(.error, log: log, "%s", message)
+      os_log(.error, log: log, "%{public}s", message)
       fatalError(message)
     }
-    
+
     guard try CoreDataPlus.isMigrationNecessary(for: sourceURL, to: targetVersion) else {
-      os_log(.info, log: log, "Migration to %s is not necessary.", "\(targetVersion.rawValue)")
+      os_log(.info, log: log, "Migration to %{public}s is not necessary.", "\(targetVersion.rawValue)")
       return
     }
-    
+
     if enableWALCheckpoint {
       os_log(.debug, log: log, "Performing a WAL checkpoint.")
       // A dead lock can occur if a NSPersistentStore with a different journaling mode
@@ -136,22 +141,22 @@ extension Migrator {
       // You need to remove it before performing a WAL checkpoint.
       try performWALCheckpointForStore(at: sourceURL, storeOptions: sourceOptions, model: sourceVersion.managedObjectModel())
     }
-    
+
     let steps = sourceVersion.migrationSteps(to: targetVersion)
-    os_log(.debug, log: log, "Number of steps: %d", steps.count)
-    
+    os_log(.debug, log: log, "Number of steps: %{public}d", steps.count)
+
     guard steps.count > 0 else { return }
-    
+
     let migrationStepsProgress = Progress(totalUnitCount: Int64(steps.count), parent: progress, pendingUnitCount: progress.totalUnitCount)
     var currentURL = sourceURL
     try steps.enumerated().forEach { (stepIndex, step) in
-      os_log(.info, log: log, "Step %d (of %d) started; %s to %s.", stepIndex + 1, steps.count, "\(step.sourceVersion)", "\(step.destinationVersion)")
+      os_log(.info, log: log, "Step %{public}d (of %{public}d) started; %{public}s to %{public}s.", stepIndex + 1, steps.count, "\(step.sourceVersion)", "\(step.destinationVersion)")
       try autoreleasepool {
         let temporaryURL = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true).appendingPathComponent(UUID().uuidString).appendingPathExtension("sqlite")
         let mappingModelMigrationProgress = Progress(totalUnitCount: Int64(step.mappingModels.count))
         migrationStepsProgress.addChild(mappingModelMigrationProgress, withPendingUnitCount: 1)
         for (mappingModelIndex, mappingModel) in step.mappingModels.enumerated() {
-          os_log(.info, log: log, "Starting migration for mapping model %d.", mappingModelIndex + 1)
+          os_log(.info, log: log, "Starting migration for mapping model %{public}d.", mappingModelIndex + 1)
           os_log(.debug, log: log, "The store at: %{public}@ will be migrated in a temporary store at: %{public}@.", currentURL as CVarArg, temporaryURL as CVarArg)
           let metadata = Metadata(sourceVersion: step.sourceVersion,
                                   sourceModel: step.sourceModel,
@@ -173,7 +178,7 @@ extension Migrator {
                                      destinationType: NSSQLiteStoreType,
                                      destinationOptions: destinationOptions)
           } catch {
-            os_log(.error, "Migration for mapping model %d failed: %{private}@", mappingModelIndex + 1, error as NSError)
+            os_log(.error, "Migration for mapping model %{public}d failed: %{private}@", mappingModelIndex + 1, error as NSError)
             throw error
           }
           let end = DispatchTime.now()
@@ -181,30 +186,30 @@ extension Migrator {
           progressReporter.markAsFinishedIfNeeded()
           let nanoseconds = end.uptimeNanoseconds - start.uptimeNanoseconds
           let timeInterval = Double(nanoseconds) / 1_000_000_000
-          os_log(.info, log: log, "Migration for mapping model %d finished in %.2f seconds.", mappingModelIndex + 1, timeInterval)
+          os_log(.info, log: log, "Migration for mapping model %{public}d finished in %{public}.2f seconds.", mappingModelIndex + 1, timeInterval)
         }
         // once the migration is done (and the store is migrated to temporaryURL)
         // the store at currentURL can be safely destroyed unless it is the
         // initial store
         if currentURL != sourceURL {
-          os_log(.debug, log: log, "Destroying store at %@.", currentURL as CVarArg)
+          os_log(.debug, log: log, "Destroying store at %{public}@.", currentURL as CVarArg)
           try NSPersistentStoreCoordinator.destroyStore(at: currentURL)
         }
         currentURL = temporaryURL
       }
-      os_log(.info, log: log, "Step %d (of %d) completed.", stepIndex + 1, steps.count)
+      os_log(.info, log: log, "Step %{public}d (of %{public}d) completed.", stepIndex + 1, steps.count)
     }
-    
+
     // move the store at currentURL to (final) destinationURL
-    os_log(.debug, log: log, "Moving the store at: %@ to final store: %{public}@.", currentURL as CVarArg, destinationURL as CVarArg)
+    os_log(.debug, log: log, "Moving the store at: %{public}@ to final store: %{public}@.", currentURL as CVarArg, destinationURL as CVarArg)
     try NSPersistentStoreCoordinator.replaceStore(at: destinationURL, withPersistentStoreFrom: currentURL)
-    
+
     // delete the store at currentURL if it's not the initial store
     if currentURL != sourceURL {
       os_log(.debug, log: log, "Destroying store at %{public}@.", currentURL as CVarArg)
       try NSPersistentStoreCoordinator.destroyStore(at: currentURL)
     }
-    
+
     // delete the initial store only if the option is set to true
     if destinationURL != sourceURL {
       os_log(.debug, log: log, "Destroying initial store at %{public}@.", sourceURL as CVarArg)
@@ -213,7 +218,7 @@ extension Migrator {
     let end = DispatchTime.now()
     let nanoseconds = end.uptimeNanoseconds - start.uptimeNanoseconds
     let timeInterval = Double(nanoseconds) / 1_000_000_000
-    os_log(.info, log: log, "Migrator has finished in %.2f seconds, final store at: %{public}@.", timeInterval, destinationURL as CVarArg)
+    os_log(.info, log: log, "Migrator has finished in %{public}.2f seconds, final store at: %{public}@.", timeInterval, destinationURL as CVarArg)
   }
 }
 
@@ -239,7 +244,7 @@ private func performWALCheckpointForStore(at storeURL: URL, storeOptions: Persis
     // https://developer.apple.com/forums/thread/118924
     options[NSPersistentHistoryTrackingKey] = [NSPersistentHistoryTrackingKey: true as NSNumber]
   }
-  
+
   let store = try persistentStoreCoordinator.addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: storeURL, options: options)
   try persistentStoreCoordinator.remove(store)
 }
@@ -252,7 +257,7 @@ extension Migrator {
     let destinationVersion: Version
     let destinationModel: NSManagedObjectModel
     let mappingModel: NSMappingModel
-    
+
     fileprivate init(sourceVersion: Version,
                      sourceModel: NSManagedObjectModel,
                      destinationVersion: Version,