+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 4.0.0 ⭐ (WIP ⚠️)
+### 4.0.0 ⭐
 
 - Added a new `Migrator` class to handle migrations (lightweight and heavyweight).
 - Added `NSDerivedAttributeDescription` utility methods.
@@ -10,6 +10,8 @@
 - Added `LightweightMigrationManger`,  a `NSMigrationManager` subclass to do *lightweight* migrations with a fake progress reporting.
 - Added `MigrationProgressReporter` to report migration progress via a `Progress` object.
 - Added a `NSManagedObjectContext` helper method to create a child context.
+- Added support for `NSPersistentStoreCoordinator` notifications payloads.
+- Many custom fetch requests now support the *affectedStores* parameter.
 - `DataTransformer` renamed as `CustomTransformer`.
 
 ### 3.0.0 ⭐